+++ projects/extension/src/environment.ts
@@ -1,3 +1,8 @@
+import westend2 from "../public/assets/westend2.json"
+import ksmcc3 from "../public/assets/ksmcc3.json"
+import polkadot from "../public/assets/polkadot.json"
+import rococo_v2_2 from "../public/assets/rococo_v2_2.json"
+
 export type StorageEntry =
   | { type: "braveSetting" }
   | { type: "database"; chainName: string }
@@ -32,6 +37,19 @@ export async function getAllActiveChains(): Promise<ExposedChainConnection[]> {
       resolve(out)
     })
   })
+}
+
+export function getDefaultBootnodes(chain: string): string[] | undefined {
+  if (chain === "polkadot") return polkadot.bootNodes
+  if (chain === "ksmcc3") return ksmcc3.bootNodes
+  if (chain === "westend2") return westend2.bootNodes
+  if (chain === "rococo_v2_2") return rococo_v2_2.bootNodes
+  return undefined
+}
+
+// Load default Bootnodes and save them to localStorage
+export async function getBootnodes(chainName: string) {
+  return await get({ type: "bootnodes", chainName })
 }
 
 export async function get<E extends StorageEntry>(